Add files via upload
The program randomly selects a number within a given range, and the user has to guess the number. After each guess, the program provides feedback:

"Too high" if the guess is greater than the number

"Too low" if the guess is less than the number

"Correct!" if the guess matches the number

The game continues until the user guesses the correct number.